+++ include/countly.hpp
@@ -25,6 +25,8 @@
 #include "countly/logger_module.hpp"
 #include "countly/storage_module_base.hpp"
 #include "countly/views_module.hpp"
+#include <countly/configuration_module.hpp>
+#include <countly/configuration_provider.hpp>
 #include <countly/crash_module.hpp>
 #include <countly/request_builder.hpp>
 #include <countly/request_module.hpp>
@@ -260,6 +262,8 @@ public:
     addEvent(event);
   }
 
+  void RecordLocation(const std::string &countryCode, const std::string &city, const std::string &gpsCoordinates, const std::string &ipAddress) override { setLocation(countryCode, city, gpsCoordinates, ipAddress); };
+
   /* Provide 'updateInterval' in seconds. */
   inline void setAutomaticSessionUpdateInterval(unsigned short updateInterval) {
     if (is_sdk_initialized) {
@@ -268,6 +272,35 @@ public:
     }
 
     configuration->sessionDuration = updateInterval;
+  }
+
+  /**
+   * Disable SDK behavior settings updates that SDK performs periodically from the server.
+   */
+  void disableSDKBehaviorSettingsUpdates() {
+    if (is_sdk_initialized) {
+      log(LogLevel::WARNING, "[Countly] disableSDKBehaviorSettingsUpdates, You can not disable SDK behavior settings updates after SDK initialization.");
+      return;
+    }
+
+    configuration->sdkBehaviorSettingsUpdatesDisabled = true;
+  }
+
+  /**
+   * Provide SDK behavior settings in JSON format string.
+   */
+  void setSDKBehaviorSettings(std::string &settings_json) {
+    if (is_sdk_initialized) {
+      log(LogLevel::WARNING, "[Countly] setSDKBehaviorSettings, You can not provide SDK behavior settings after SDK initialization.");
+      return;
+    }
+
+    if(settings_json.empty()) {
+      log(LogLevel::WARNING, "[Countly] setSDKBehaviorSettings, Provided SDK behavior settings is empty.");
+      return;
+    }
+
+    configuration->sdkBehaviorSettings = settings_json;
   }
 
 #ifdef COUNTLY_BUILD_TESTS
@@ -348,6 +381,7 @@ private:
   std::shared_ptr<cly::RequestBuilder> requestBuilder;
   std::shared_ptr<cly::RequestModule> requestModule;
   std::shared_ptr<cly::StorageModuleBase> storageModule;
+  std::shared_ptr<cly::ConfigurationModule> configurationModule;
   std::shared_ptr<std::mutex> mutex = std::make_shared<std::mutex>();
 
   bool is_queue_being_processed = false;