+++ src/crash_module.cpp
@@ -16,6 +16,7 @@ public:
   std::shared_ptr<LoggerModule> _logger;
   std::shared_ptr<RequestModule> _requestModule;
   std::shared_ptr<std::mutex> _mutex;
+  std::weak_ptr<ConfigurationProvider> _configProvider;
   CrashModuleImpl(std::shared_ptr<CountlyConfiguration> config, std::shared_ptr<LoggerModule> logger, std::shared_ptr<RequestModule> requestModule, std::shared_ptr<std::mutex> mutex) : _configuration(config), _logger(logger), _requestModule(requestModule), _mutex(mutex) {}
 
   // destructor to reset logger
@@ -49,6 +50,16 @@ void CrashModule::addBreadcrumb(const std::string &value) {
 void CrashModule::recordException(const std::string &title, const std::string &stackTrace, const bool fatal, const std::map<std::string, std::string> &crashMetrics, const std::map<std::string, std::string> &segmentation) {
 
   impl->_logger->log(LogLevel::INFO, cly::utils::format_string("[CrashModule] recordException: title = %s, stackTrace = %s", title.c_str(), stackTrace.c_str()));
+
+  if (std::shared_ptr<ConfigurationProvider> config = impl->_configProvider.lock()) {
+    if (config->isCrashReportingEnabled() == false) {
+      impl->_logger->log(LogLevel::DEBUG, "[CrashModule] recordException: Crash reporting is disabled. Not recording exception.");
+      return;
+    }
+  } else {
+    impl->_logger->log(LogLevel::WARNING, "[CrashModule] recordException: ConfigurationProvider unavailable.");
+    return;
+  }
 
   if (title.empty()) {
     impl->_logger->log(LogLevel::WARNING, "[CrashModule] recordException : The parameter 'title' can't be empty");
@@ -93,5 +104,7 @@ void CrashModule::recordException(const std::string &title, const std::string &s
   // unlock mutex
   impl->_mutex->unlock();
 }
+
+void CrashModule::setConfigurationProvider(std::weak_ptr<ConfigurationProvider> provider) { impl->_configProvider = std::move(provider); }
 
 } // namespace cly